Multi-Step-Agent/Simple_Agent/pyproject.toml
[project]
name = "simple-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-mcp-adapters>=0.1.7",
    "langchain-openai>=0.3.24",
    "langgraph>=0.4.8",
    "python-dotenv>=1.1.0",
]
